Refactoring structure
