+++ requirements.txt
@@ -1,0 +1,1 @@
+aiogram==3.21.0
